Adjust snapshot unit tests for Windows additions
